Updated return value for getCountOfInsurancePolicies

+++ api/src/main/java/org/openmrs/module/mohbilling/db/BillingDAO.java
@@ -742,5 +742,5 @@ public interface BillingDAO {
 
 	List<InsurancePolicy> getInsurancePoliciesByPagination(int offset, int limit);
 
-	int getInsurancePolicyCount();
+	long getInsurancePolicyCount();
 }
